[skip ci]Update get_segment_distribution function
